feat: 팀 멤버 디테일 모달 구현 #71
- 팀 멤버 카드 컴포넌트 파일 분리
- 팀 멤버 디테일 모달 구현
